Moving config() to it's own Class to have static values
currently config() reads the config file each time it's called. It's
only called a handful of times, so not crazy overhead but still
something that makes me itch...

Moved `config()` into a class `Config` which checks at init to see if
the config file has been read - if not, it reads/parses it. After class
instantiation, callers can then access it's member values.